+++ R/rule_css.R
@@ -19,12 +19,9 @@ rule_css <- function(x, columns, expression,
                      na.value = "",
                      lockcells = FALSE) {
   columnsquo <- rlang::enquo(columns)
-  helpers <- tidyselect::vars_select_helpers
-  columnsquo_bur <- rlang::env_bury(columnsquo, !!! helpers)
-
   expr <- rlang::enquo(expression)
 
-  rule <- structure(list(columns = columnsquo_bur,
+  rule <- structure(list(columns = columnsquo,
                          expression = expr,
                          css_field = force(css_field),
                          na.value = force(na.value),
@@ -36,18 +33,18 @@ rule_css <- function(x, columns, expression,
 }
 
 rule_to_cf_field.rule_css <- function(rule, xfiltered, xview, ...) {
-  columns <- do.call(tidyselect::vars_select, c(list(colnames(xview)), rule[["columns"]]))
+  columns <- tidyselect::eval_select(expr = rule[["columns"]], data = xview)
   if (length(columns) == 0) {
     return(NULL)
   }
   if (rlang::quo_is_missing(rule[["expression"]])) {
     if (length(columns) > 1) {
       warning("rule_css applied to multiple columns, using column ",
-              columns[1], " values as expression. In the future this behaviour will change,",
+              names(columns)[1], " values as expression. In the future this behaviour will change,",
               "please use a explicit expression instead.",
               call. = FALSE)
     }
-    rule[["expression"]] <- as.symbol(as.name(columns[1]))
+    rule[["expression"]] <- rlang::sym(names(columns)[1])
   }
   css_values <- rlang::eval_tidy(rule[["expression"]], data = xfiltered)
   stopifnot(identical(length(css_values), nrow(xview)))