+++ R/rule_text_bold.R
@@ -16,12 +16,9 @@ rule_text_bold <- function(x, columns, expression,
                            na.bold = FALSE,
                            lockcells = FALSE) {
   columnsquo <- rlang::enquo(columns)
-  helpers <- tidyselect::vars_select_helpers
-  columnsquo_bur <- rlang::env_bury(columnsquo, !!! helpers)
-
   expr <- rlang::enquo(expression)
 
-  rule <- structure(list(columns = columnsquo_bur,
+  rule <- structure(list(columns = columnsquo,
                          expression = expr,
                          na.value = force(na.bold),
                          lockcells = force(lockcells)),
@@ -32,18 +29,18 @@ rule_text_bold <- function(x, columns, expression,
 }
 
 rule_to_cf_field.rule_text_bold <- function(rule, xfiltered, xview, ...) {
-  columns <- do.call(tidyselect::vars_select, c(list(colnames(xview)), rule[["columns"]]))
+  columns <- tidyselect::eval_select(expr = rule[["columns"]], data = xview)
   if (length(columns) == 0) {
     return(NULL)
   }
   if (rlang::quo_is_missing(rule[["expression"]])) {
     if (length(columns) > 1) {
       warning("rule_text_bold applied to multiple columns, using column ",
-              columns[1], " values as expression. In the future this behaviour will change,",
+              names(columns)[1], " values as expression. In the future this behaviour will change,",
               "please use a explicit expression instead.",
               call. = FALSE)
     }
-    rule[["expression"]] <- as.symbol(as.name(columns[1]))
+    rule[["expression"]] <- rlang::sym(names(columns)[1])
   }
   bold_or_not <- rlang::eval_tidy(rule[["expression"]], data = xfiltered)
   stopifnot(identical(length(bold_or_not), nrow(xview)))